+++ helpers/index.js
@@ -1,8 +1,12 @@
 module.exports = {
     addChoicesToSelect: require('./addChoicesToSelect').addChoicesToSelect,
     batchProcess: require('./batchProcess').batchProcess,
+    checkFetchStatus: require('./checkFetchStatus').checkFetchStatus,
     checkPermissions: require('./checkPermissions').checkPermissions,
-    formatQuery: require('./formatQuery').formatQuery,
+    createRequiredFields: require('./createRequiredFields').createRequiredFields,
+    parseResponse: require('./parseResponse').parseResponse,
+    setSettingValues: require('./setSettingValues').setSettingValues,
+    settleAllPromises: require('./settleAllPromises').settleAllPromises,
     wait: require('./wait').wait,
     validateTable: require('./validateTable').validateTable,
 }