+++ src/d2l.js
@@ -4,52 +4,109 @@
  * HCat API - D2L endpoints
  * 
  * Supported endpoints:
- *      Course:
- *          - importCourse()
- *          - getImportCourseJobStatus()
- *          - getCourseListing()
- *      OrgUnit:
- *          - getOrgInfo()
- *      Users:
- *          - createUser()
- *          - fetchUser()
- *          - updateUserPassword()
- *          - createUserEnrollment()
- *          - deleteUserEnrollment()
+ *      Awards:
+ *          - createAward()
+ *          - createAwardAssociation()
+ *          - getAvailableAwards()
+ *          - issueAward()
  *      Content:
  *          - getContent()
  *          - createContent()
  *          - createDiscussionTopic()
+ *          - updateTopic()
+ *      Course:
+ *          - getCourseListing()
+ *          - getImportCourseJobStatus()
+ *          - importCourse()
+ *      Misc:
+ *          - whoAmI()
+ *      OrgUnit:
+ *          - addReleaseCondition()
+ *          - getOrgInfo()
+ *      Users:
+ *          - createUser()
+ *          - fetchUser()
+ *          - modifyUserEnrollment()
+ *          - updatePreferredNames()
+ *          - updateUserPassword()
  */
-
 class D2LApi {
     constructor(hcat) {
-        this.endpoint = "/d2l";
+        this.endpoint = "/api/d2l";
         this.hcat = hcat;
     }
 
-    /**
-     * Create Import course Copy Job.
-     * 
-     * @param {String} apiKey - LMS auth token.
-     * @param {String|number} sourceId - ID for the source course.
-     * @param {String|number} targetId - ID for the target course.
-     * @param {Number} offset - Number of days to offset course dates.
-     * @return {Object} - JobToken id to use for checking status of job.
-     */
-    async importCourse(sourceId, targetId, offset = null) {
-        // Construct endpoint with Target Course
-        let endpoint = `${this.endpoint}/${targetId}/import`;
+    /*--------------------------------------------------------------
+    # AWARDS
+    --------------------------------------------------------------*/
+    async createAward(payload) {
+        let endpoint = `${this.endpoint}/awards`;
 
-        // Construct payload with Source Course Id
-        const payload = {
-            "SourceOrgUnitId": sourceId,
-            "Components": null,
-            "CallbackUrl": null,
-            "DaysToOffsetDates": offset,
+        let response = await this.hcat.fetchWrapper({method: "POST", endpoint: endpoint, payload: payload});
+        return response;
+    }
+    
+    async createAwardAssociation(orgUnitId, payload) {
+        let endpoint = `${this.endpoint}/${orgUnitId}/associations`;
+        let response = await this.hcat.fetchWrapper({method: "POST", endpoint: endpoint, payload: payload});
+        return response;
+    }
+
+    async getAvailableAwards(orgUnitId) {
+        let endpoint = `${this.endpoint}/${orgUnitId}/associations`;
+        return this.hcat.fetchWrapper({endpoint: endpoint});
+    }
+
+    async issueAward(orgUnitId, payload) {
+        let endpoint = `${this.endpoint}/${orgUnitId}/issued`;
+        return this.hcat.fetchWrapper({method: "POST", endpoint: endpoint, payload: payload});
+    }
+
+    /*--------------------------------------------------------------
+    # CONTENT
+    --------------------------------------------------------------*/
+    async getContent(orgUnitId, type, id, file=false) {
+        let endpoint = `${this.endpoint}/${orgUnitId}/${type}/${id}`;
+
+        if (file) {
+            endpoint += "/file";
         }
 
-        return this.hcat.fetchWrapper("POST", endpoint, payload);
+        return this.hcat.fetchWrapper({endpoint: endpoint});
+    }
+
+    async createContent(payload) {
+        let endpoint = `${this.endpoint}/${payload.OrgUnitProperties.Identifier}/${payload.type}`;
+        return this.hcat.fetchWrapper({method: "POST", endpoint: endpoint, payload: payload});
+    }
+    
+    async createDiscussionTopic(payload) {
+        let endpoint = `${this.endpoint}/${payload.OrgUnitProperties.Identifier}/discussions/${payload.parentModule.ForumId}/topics`;
+        return this.hcat.fetchWrapper({method: "POST", endpoint: endpoint, payload: payload});
+    }
+    
+    async updateTopic(payload, file=false) {
+        let endpoint = `${this.endpoint}/${payload.OrgUnitProperties.Identifier}/topics/${payload.Id}`;
+
+        if (file) {
+            endpoint += "/file";
+        }
+
+        return this.hcat.fetchWrapper({method: "PUT", endpoint: endpoint, payload: payload});
+    }
+
+    /*--------------------------------------------------------------
+    # COURSE
+    --------------------------------------------------------------*/
+    async getCourseListing({orgUnitId=null, queryParams = {}}) {
+        let endpoint = `${this.endpoint}/orgstructure/`;
+
+        // Request specific course if ID provided        
+        if (orgUnitId) {
+            endpoint += orgUnitId;
+        }
+
+        return this.hcat.fetchWrapper({endpoint: endpoint, queryParams: queryParams});
     }
 
     /**
@@ -62,7 +119,6 @@ class D2LApi {
      * It is the job of the caller to respect the API resources and restrict
      * calls after a number of iterations/times.
      * 
-     * @param {String} apiKey - LMS auth token.
      * @param {String|number} targetId - ID for the target course.
      * @param {String} jobId - The JobToken for the copy.
      * @returns {String} The current status of the copy job.
@@ -70,78 +126,90 @@ class D2LApi {
     async getImportCourseJobStatus(targetId, jobId) {
         let endpoint = `${this.endpoint}/${targetId}/import/${jobId}`;
 
-        let response = await this.hcat.fetchWrapper("GET", endpoint);
+        let response = await this.hcat.fetchWrapper({endpoint: endpoint});
         return response.Status;
     }
 
+    /**
+         * Create Import course Copy Job.
+         * 
+         * @param {String|number} sourceId - ID for the source course.
+         * @param {String|number} targetId - ID for the target course.
+         * @param {Number} offset - Number of days to offset course dates.
+         * @return {Object} - JobToken id to use for checking status of job.
+         */
+    async importCourse(sourceId, targetId, offset = null) {
+        let endpoint = `${this.endpoint}/${targetId}/import`;
 
-    async getCourseListing({orgUnitId=null, queryParams = {}}) {
-        let endpoint = `/d2l/orgstructure/`;
-
-        // Request specific course if ID provided        
-        if (orgUnitId) {
-            endpoint += orgUnitId;
+        // Construct payload with Source Course Id
+        const payload = {
+            "SourceOrgUnitId": sourceId,
+            "Components": null,
+            "CallbackUrl": null,
+            "DaysToOffsetDates": offset,
         }
 
-        return this.hcat.fetchWrapper("GET", endpoint, null, queryParams);
+        return this.hcat.fetchWrapper({method: "POST", endpoint: endpoint, payload: payload});
+    }
+
+    /*--------------------------------------------------------------
+    # MISC
+    --------------------------------------------------------------*/
+    async whoAmI() {
+        let endpoint = `${this.endpoint}/users/whoami`;
+        return this.hcat.fetchWrapper({endpoint: endpoint});
+    }
+    
+    /*--------------------------------------------------------------
+    # ORG UNIT
+    --------------------------------------------------------------*/
+    async addReleaseCondition(orgUnitId, type, typeId, conditionType) {
+        let endpoint = `${this.endpoint}/${orgUnitId}/conditionalRelease/${type}/${typeId}`;
+        const payload = {
+            "Operator": "All",
+            "Type": conditionType,
+        }
+        return this.hcat.fetchWrapper({method: "PUT", endpoint: endpoint, payload: payload});
     }
 
     async getOrgInfo(orgUnitId, target) {
-        let endpoint = `/d2l/${orgUnitId}/${target}`;
-        return this.hcat.fetchWrapper("GET", endpoint);
+        let endpoint = `${this.endpoint}/${orgUnitId}/${target}`;
+        return this.hcat.fetchWrapper({endpoint: endpoint});
     }
-
+    
+    /*--------------------------------------------------------------
+    # USERS
+    --------------------------------------------------------------*/
     async createUser(payload) {
-        let endpoint = `/d2l/users`;
-        return this.hcat.fetchWrapper("POST", endpoint, payload);
+        let endpoint = `${this.endpoint}/users`;
+        return this.hcat.fetchWrapper({method: "POST", endpoint: endpoint, payload: payload});
     }
 
     async fetchUser(queryParams) {
-        let endpoint = `/d2l/users`;
-        return this.hcat.fetchWrapper("GET", endpoint, null, queryParams);
+        let endpoint = `${this.endpoint}/users`;
+        return this.hcat.fetchWrapper({endpoint: endpoint, queryParams: queryParams});
+    }
+
+    async modifyUserEnrollment(action, orgUnitId, user) {
+        let endpoint = `${this.endpoint}/${orgUnitId}/enroll/${user.fields.Identifier}`;
+
+        const method = (action === "enroll") ? "POST" : "DELETE";
+        const payload = (action === "enroll") ? {"RoleId": user.fields.RoleId} : null;
+
+        return this.hcat.fetchWrapper({method: method, endpoint: endpoint, payload: payload});
+    }
+
+    async updatePreferredNames(userId, payload) {
+        let endpoint = `${this.endpoint}/users/${userId}/names`;
+        return this.hcat.fetchWrapper({method: "PUT", endpoint: endpoint, payload: payload});
     }
 
     async updateUserPassword(userId, password) {
-        let endpoint = `/d2l/users/${userId}/password`;
+        let endpoint = `${this.endpoint}/users/${userId}/password`;
         let payload = {"Password": password}
 
-        return this.hcat.fetchWrapper("PUT", endpoint, payload);
-    }
-
-    async createUserEnrollment(orgUnitId, user) {
-        let endpoint = `/d2l/${orgUnitId}/enroll/${user.fields.Identifier}`;
-
-        const payload = {
-            "RoleId": user.fields.RoleId,
-        }
-
-        return this.hcat.fetchWrapper("POST", endpoint, payload);
-    }
-
-    async deleteUserEnrollment(orgUnitId, userId) {
-        let endpoint = `/d2l/${orgUnitId}/enroll/${userId}`;
-        return this.hcat.fetchWrapper("DELETE", endpoint);
-    }
-
-    async getContent(orgUnitId, type, id, file=false) {
-        let endpoint = `/d2l/${orgUnitId}/${type}/${id}`;
-
-        if (file) {
-            endpoint += "/file";
-        }
-
-        return this.hcat.fetchWrapper("GET", endpoint);
-    }
-
-    async createContent(payload) {
-        let endpoint = `/d2l/${payload.OrgUnitProperties.Identifier}/${payload.type}`;
-        return this.hcat.fetchWrapper("POST", endpoint, payload);
-    }
-
-    async createDiscussionTopic(payload) {
-        let endpoint = `/d2l/${payload.OrgUnitProperties.Identifier}/discussions/${payload.parentModule.ForumId}/topics`;
-        return this.hcat.fetchWrapper("POST", endpoint, payload);
-    }
+        return this.hcat.fetchWrapper({method: "PUT", endpoint: endpoint, payload: payload});
+    }    
 }
 
 module.exports = D2LApi;